Refactor `costs` package to be more concise; separate based on domain
Previously split based on costs, but it becomes more cumbersome to write new costs. Additionally, with this setup we need a uniform data structure as an interface for different costs, which the template CostTerm cannot provide. The solution is to merge all the costs together in a manager, and have delegates to handle computation of domain-specific variables such as vehicular slip angle and contouring error